+++ app/src/main/java/com/team2/chitchat/data/usecase/GetMessagesUseCase.kt
@@ -1,7 +1,7 @@
 package com.team2.chitchat.data.usecase
 
 import com.team2.chitchat.data.domain.model.messages.GetMessagesModel
-import com.team2.chitchat.data.repository.remote.backend.DataProvider
+import com.team2.chitchat.data.repository.DataProvider
 import com.team2.chitchat.data.repository.remote.response.BaseResponse
 import kotlinx.coroutines.flow.Flow
 import javax.inject.Inject
